Move BrowserAndLoadingPanel.java
